backend/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.0
